+++ playground/app/pages/cached.vue
@@ -1,0 +1,7 @@
+<script setup>
+const { data } = await useFetch('/api/cached')
+</script>
+
+<template>
+  <pre>{{ data }}</pre>
+</template>