+++ src/runtime/base/server/utils/hooks.ts
@@ -1,5 +1,4 @@
 import { createHooks } from 'hookable'
-import { useRuntimeConfig } from '#imports'
 
 export interface HubHooks {
   'bindings:ready': () => void
@@ -28,8 +27,7 @@ export const hubHooks = createHooks<HubHooks>()
  * @see https://hub.nuxt.com/docs/recipes/hooks#onhubready
  */
 export function onHubReady(cb: HubHooks['bindings:ready']) {
-  const hub = useRuntimeConfig().hub
-  if (import.meta.dev && !hub.remote) {
+  if (import.meta.dev) {
     return hubHooks.hookOnce('bindings:ready', cb)
   }
   cb()