+++ src/runtime/cache/server/api/_hub/cache/[...key].delete.ts
@@ -1,7 +1,6 @@
 import { eventHandler, getRouterParam, createError, sendNoContent } from 'h3'
 import { requireNuxtHubAuthorization } from '../../../../../utils/auth'
 import { requireNuxtHubFeature } from '../../../../../utils/features'
-// @ts-expect-error useStorage not yet typed
 import { useStorage } from '#imports'
 
 export default eventHandler(async (event) => {