+++ src/types/blob.ts
@@ -210,3 +210,47 @@ export interface BlobListResult {
    */
   folders?: string[]
 }
+
+export interface BlobCredentialsOptions {
+  /**
+   * The permission of the credentials.
+   * @default 'admin-read-write'
+   */
+  permission?: 'admin-read-write' | 'admin-read-only' | 'object-read-write' | 'object-read-only'
+  /**
+   * The ttl of the credentials in seconds.
+   * @default 900
+   */
+  ttl?: number
+  /**
+   * The prefixes to scope the credentials to.
+   */
+  prefixes?: string[]
+  /**
+   * The pathnames to scope the credentials to.
+   */
+  pathnames?: string[]
+}
+
+export interface BlobCredentials {
+  /**
+   * The Cloudflare account id
+   */
+  accountId: string
+  /**
+   * The Cloudflare R2 bucket name
+   */
+  bucketName: string
+  /**
+   * The access key id for the R2 bucket
+   */
+  accessKeyId: string
+  /**
+   * The secret access key for the R2 bucket
+   */
+  secretAccessKey: string
+  /**
+   * The session token for the R2 bucket
+   */
+  sessionToken: string
+}